+++ examples/hello.rs
@@ -2,7 +2,7 @@ use fibra::*;
 
 #[tokio::main]
 async fn main() -> FibraResult<()> {
-    // create a main router with a logger
+    // create a main router with a predefined logger
     let mut app = Fibra::default();
     app.mount(addon::Logger::default());
     app.route("/", app_root)?;
@@ -13,42 +13,41 @@ async fn main() -> FibraResult<()> {
     api.limit().host("api.*");
     api.route("/", api_root)?;
 
-    // create api version 1 router
+    // create api/v1's subrouter
     let v1 = api.group("/v1")?;
     v1.route("/", v1_root)?;
     v1.route("/user", v1_user)?;
 
-    // create api version 2 router
+    // create api/v2's subrouter
     let v2 = api.group("/v2")?;
     v2.route("/", v2_root)?;
     v2.route("/user", v2_user)?;
 
-    // bind on two ports and run the server
-    app.bind("0.0.0.0:3000")?;
-    app.bind("0.0.0.0:3333")?;
+    // bind on a port and run the server
+    app.bind("0.0.0.0:8000")?;
     app.run().await
 }
 
-async fn app_root(_ctx: Context) -> FibraResult<Response<Body>> {
+async fn app_root(_ctx: Context) -> FibraResult<Response> {
     Ok((StatusCode::OK, "It Works!").into_response())
 }
 
-async fn api_root(_ctx: Context) -> FibraResult<Response<Body>> {
-    todo!()
+async fn api_root(_ctx: Context) -> FibraResult<Response> {
+    Ok((StatusCode::OK, "Api Root").into_response())
 }
 
-async fn v1_root(_ctx: Context) -> FibraResult<Response<Body>> {
-    todo!()
+async fn v1_root(_ctx: Context) -> FibraResult<Response> {
+    Ok((StatusCode::OK, "V1 Root").into_response())
 }
 
-async fn v1_user(_ctx: Context) -> FibraResult<Response<Body>> {
-    todo!()
+async fn v1_user(_ctx: Context) -> FibraResult<Response> {
+    Ok((StatusCode::OK, "User1").into_response())
 }
 
-async fn v2_root(_ctx: Context) -> FibraResult<Response<Body>> {
-    todo!()
+async fn v2_root(_ctx: Context) -> FibraResult<Response> {
+    Ok((StatusCode::OK, "V2 Root").into_response())
 }
 
-async fn v2_user(_ctx: Context) -> FibraResult<Response<Body>> {
-    todo!()
+async fn v2_user(_ctx: Context) -> FibraResult<Response> {
+    Ok((StatusCode::OK, "User2").into_response())
 }