+++ src/addon/normalize.rs
@@ -5,7 +5,7 @@ pub struct Normalize;
 
 #[async_trait]
 impl Handler for Normalize {
-    async fn handle(&self, _ctx: Context) -> FibraResult<Response<Body>> {
+    async fn handle(&self, _ctx: Context) -> FibraResult<Response> {
         todo!()
     }
 }